Update Dockerfile to use opam 2
A big difference with the previous Dockerfile is that this container
holds the new learn-ocaml program, not the full web-app: it should be
run with a local exercises repository.

This could be used as base for a container including the exercises
though, possibly using Docker multi-stage.

Build the image using `make docker-image`