fix Mathjax URL
apparently their CDN has been discontinued. See
https://www.mathjax.org/cdn-shutting-down/

Note that relying on a CDN is NOT a good option if we want to be able to
run isolated servers (eg on a local network). We should embed a stripped
version of Mathjax for that purpose (the page linked above has some tips
about this, to avoid including too much useless weight)