Client improvements (fixed)
This version of the patch includes a fix to the linking bug that broke
the web-app.

This adds a configuration file to the client (either
`~/.config/learn-ocaml/client.json`, or a dir-local
`./.learnocaml-client` file), and an init procedure for providing or
generating the config. The usage then become transparent, without having
to define variables for the server or token.

This enforces the user to have a token before grading, which is not a
technical requirement but is, in my opinion, a better behaviour: it
ensures the client-side is well configured and submits the solution at
every grading.

Many options have been added as well, e.g. to retrieve the current
student solutions to a local directory.