test_lib.mli: Add another way to build ['a Ty.ty] terms
Use case: the type of a function to grade depends on a type [t] that
should be defined in the student's code itself.

If that type is undefined, [%ty: Code.t] will trigger a syntax error,
while (ty_of "Code.t") will fail locally only, and let the other tests
pass.