Test_functions_generic: Expose 2 helper functions & Add an infix
These helper functions are necessary to get a «'a Ty.ty» from a «prot»
in order to avoid recomputing the ty when a (more-informative) prot
has already be constructed using last_ty and arg_ty.  This can be
useful for example if we want to use test_function and lookup_student.

The example below also needs the infix notation (==>) that is added in
this commit.

  let prot = arg_ty [%ty: bool] @@ last_ty [%ty: bool] [%ty: bool] in
  test_function prot (lookup_student (ty_of_prot prot) "foo")
    [true @:!! true ==> true;
     true @:!! false ==> false;
     false @:!! true ==> true;
     false @:!! false ==> true]