Configurable welcome screen announce and nickname
This makes the "nickname" placeholder and the login page announce
configurable through `<webroot>/js/learnocaml-config.js`, using the
fields `txtNickname` and `txtLoginWelcome`.

At the moment, this file is overwritten on `learn-ocaml build` and needs
to be edited by hand, though. Options could be added to the command, but
that won't scale if we want more customisable fields.

Also note that specifying these fields will disable use of localised
strings on the clients.

Closes #123