Translate messages from the toplevel
Warning: this needs a fix to js_of_ocaml to ensure utf8 characters from
ocaml are correctly printed, otherwise they get mangled and no
translation is probably better.

The included `learn-ocaml.opam.locked` file includes a pin-depends to
js_of_ocaml with the required patch, so that should be OK for the docker
containers though.