Learnocaml_data: refactor Server configuration
The previous API was unclear; I couldn't understand the role of
enc_init without reading the client code, and it was not clear how to
extend server configuration with new fields. After the refactoring,
it's clear that adding a new field requires thinking about whether it
should be part of the preconfig or not -- and it could be optional in
the preconfig and mandatory in the config, etc.