[minor] refactor Cmdliner usage
The goal is to make it as simple as possible to add new fields to
configuration records (Builder.t, Server.t, etc).

The command-line-interface parts of learnocaml_main entangle the
construction of proper configuration records with unpleasant
global-mutable-variable settings. This PR splits each affected 'term'
value into smaller parts, with one purely in charge of building the
configuration record, and another dealing with the global mutation.