QoL: reset the toplevel every time before evaluating code from the editor
You might lose what you manually typed directly in the toplevel pane, but that
shouldn't be a problem, and most importantly it will avoid already defined ids
to clutter the environment.

These often cause student confusion with cases like "type t instead of t", or a
missing `rec` keyword that compiles because using an older definition.