repo: fix repo generation when an exercise has multiple subdirectories
When an exercise is copied to the destination directory, its
subdirectories (and their parents) are copied concurrently.  This is
not an issue when there is only one subdirectory to copy, but not when
there is multiple subdirectories.  In this case, for each subdirectory
of the exercise, `mkdir_p' will check if the parent directory exists.
As it does not, they will all attempt to create the parent.  One of
those will succeed, and the rest will fail with EEXIST.

This solves this issue by creating the exercise's destination
directory before creating subdirectories.

This fixes #305.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>