app/index_main: disable the sync button by default
When learn-ocaml is statically deployed, the buttons related to the
token cannot really be used.  Disable them when they are created, then
enable them once the token is validated.

As `sync_button_state' is unused and not fit for this purpose, it is
replaced by a button group, `sync_button_group'.  The buttons are
associated to this group when they are created.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>