app/index_main: select the correct URL for the deployment mode
If learn-ocaml is deployed statically, the URL of the editor should be
exercise.html, and the URL of the playground must be playground.html.
But when deployed with the server, it can use the clean URL.  To
detect it, we check the token: if it does not exists, it means the
application is deployed without server.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>