/launch: let the server fill the form in launch.html itself
For the association to an existing account to work, the user id given
by the LTI application should be forwarded to the page doing the work.

Let the csrf token be filled too, while we're at it.

The used-id field is not secured, an attacker could associate an
account that doesn't exist yet with a specific token.  An HMAC could
fix this issue.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>