learnocaml_index_main: read the token in the cookies if it is set
As the front-end cannot access to POST parameters, one reliable way to
pass parameters from the back-end to the front-end is by using
cookies.  The page /launch and /launch/login both set the cookie
`token' if the token is correct, so instead of asking the user for its
token once again, try to retrieve it from the cookies before asking it
to the user.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>