learnocaml_index_main: don't show the token if it can't be used to login
This adds a new key in the local storage to store this information.
If the login happened by registering with an email/password pair, by
login in with an email/password pair, or with a cookie (indicating a
login through LTI), the underlying token cannot be used to login
directly, so the stored value will be false.  In this case, the token
will not be displayed when login out, and the button will be disabled.

If the user logged in by inputing the token, and that it succedeed,
the token can be displayed to the user.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>