learnocaml_common: hide the dialog box first, then call the callback
When creating dialog boxes with buttons, the provided callback is
called before hiding the box.

This is not a problem most of the time, but it becomes one when the
callback implies Lwt threads and other dialog boxes.  Take
`ask_string', for instance: its validation button wakes up another Lwt
thread, and if it happens to create another dialog box (which reuses
the same basic <div> to show its components as the one created by
`ask_string'), when the button's callback is awoken again, it will
hide the newly created box.

Such a sequence didn't exist until recently, with 0296f6c (fix: Make
email validation check more strict and uniform, 2020-08-27).  It
introduces a sequence like this:

  ask_string … >>= fun str ->
  if is_valid str then Lwt.return_some str
  else begin
    alert …;
    Lwt.return_none
  end

Here, the callback of the "Ok" button will wake up `ask_string', and
if alert is called, it will show up for a fraction of second before
the button's callback is in charge again and hide this dialog.

This fix this issue by hiding the dialog box first, then calling the
callback.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>