server_config.json: allow bool values to be specified with strings
This allow system administrators to specify boolean
values (ie. `use_moodle' and `use_passwd') with strings (ie. "true" or
"false") instead of "genuine" booleans (ie. `true' or `false') to be
more graceful with mistakes.

Note that while no string should cause a hard crash, if it is not
strictly equal to "true", the value will be interpreted as false (even
"True"), since this is the behaviour of `bool_of_string_opt'.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>