config: add an `enableMoodle' value to `learnocaml_config'
This adds an `enableMoodle' value to `learnocaml_config' to allow the
front-end to know about the status of Moodle.  This will be useful
later when a message relying on this will be added.

Furthermore, the `learnocaml_config' in learnocaml_index_main.ml has
been removed, and `get_opt' is moved to learnocaml_config.ml.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>