Provide lib to compile grader helper libraries
and facility to link them during the `build` step.

NOTE: the helper library is going to be included in every exercise. A
lighter approach could be to keep loading it separately, e.g. after
loading the cma/js file from a directory holding static content on the
server (and removing the `.cma` from the compilation line in
`precompile_exercise.ml`).

This will probably fit well once we include such a mechanism for loading
custom libraries as the prelude to exercises, the main difference being
that the latter will also need the `cmi` files.