build: update pinned package versions
Several versions of packages mentioned in the lockfiles before this
commit do not exist on opam anymore. This causes a build failure on a
fresh clone of this repo in docker, which uses the package versions
pinned in the lockfiles.

To find these package versions, I simply deleted the lockfiles,
installed all dependencies in a fresh OCaml 5.1.1 switch, and used `opam
lock`. This process resulted in version 5.9.1 for
js_of_ocaml{,-toplevel,-compiler,...} on which Learn-OCaml's toplevel
fails to initialize, making the webapp unusable. I then manually lowered
the pinned version of js_of_ocaml back to the previously pinned version,
5.8.2.